feat: add creator stats store
