+++ apps/web/pages/api/creator-stats.ts
@@ -1,19 +1,8 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import fs from 'fs';
 import path from 'path';
+import { CreatorStatsStore } from '../../lib/creatorStatsStore';
 
-interface Store {
-  [pubkey: string]: any;
-}
-
-function loadStore(): Store {
-  try {
-    const p = path.join(process.cwd(), 'scripts', 'creator-stats.json');
-    return JSON.parse(fs.readFileSync(p, 'utf8'));
-  } catch {
-    return {};
-  }
-}
+const store = new CreatorStatsStore(path.join(process.cwd(), 'scripts', 'creator-stats.json'));
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   const pubkey = req.query.pubkey as string;
@@ -26,7 +15,11 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
     res.status(403).json({ error: 'forbidden' });
     return;
   }
-  const store = loadStore();
-  const stats = store[pubkey] || { totals: { views: 0, zapsSats: 0, comments: 0, followerDelta: 0, revenueAud: 0 }, dailySeries: [] };
+  const data = store.read();
+  const stats =
+    data[pubkey] || {
+      totals: { views: 0, zapsSats: 0, comments: 0, followerDelta: 0, revenueAud: 0 },
+      dailySeries: [],
+    };
   res.status(200).json(stats);
 }
